Added CustomUserDetails to retrieve the user's ID in the principal. Implemented a custom version of @WithMockUser for testing purposes - @WithMockUserCustomer, along with its security context factory WithMockCustomUserSecurityContextFactory, to access principal.id for frontend testing (due to the navbar that contains principal.id). Also fixed a bug related to deleting a user without an avatar.

+++ src/main/java/com/drunar/coincorner/service/ImageService.java
@@ -62,14 +62,18 @@ public class ImageService {
 
     @SneakyThrows
     public boolean delete(String imagePath) {
-        Path fullImagePath = Path.of(bucket, imagePath);
+        if (imagePath != null && !imagePath.isEmpty()) {
+            Path fullImagePath = Path.of(bucket, imagePath);
 
-        if (Files.exists(fullImagePath)) {
-            try {
-                Files.delete(fullImagePath);
-                return true;
-            } catch (IOException e) {
-                e.printStackTrace();
+            if (Files.exists(fullImagePath)) {
+                try {
+                    Files.delete(fullImagePath);
+                    return true;
+                } catch (IOException e) {
+                    e.printStackTrace();
+                    return false;
+                }
+            } else {
                 return false;
             }
         } else {
